% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra2.R
\name{dijkstra2}
\alias{dijkstra2}
\title{Dijkstra Algorithm modified}
\usage{
dijkstra2(graph, init_node)
}
\arguments{
\item{graph}{Dataframe}

\item{init_node}{Scalar}
}
\value{
Numeric vector representing the shortest distance form 
        init_node to all the other nodes in graph
}
\description{
This function finds the shortest distance from a source 
             node to all the other nodes in a graph
}
\details{
Advanced Programming in R - Lab 3
Dijkstra algorithm modified
}
\examples{
dijkstra2(wiki_graph2, "a")
}
