% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm for greatest common divisor}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{\itemize{
\item A numeric scalar.
}}

\item{b}{\itemize{
\item A numeric scalar.
}}
}
\value{
The greatest common divisor of a and b as a numeric scalar.
}
\description{
Euclidean algorithm for greatest common divisor
}
\details{
Implements the Euclidean algorithm:
repeatedly replace the larger number with its remainder when divided
by the smaller number until the remainder is zero.
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
