% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidian.R
\name{euclidian}
\alias{euclidian}
\title{\code{euclidian} returns the lowest common denominator between two numbers.}
\usage{
euclidian(a, b)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The lowest common denominator of \code{x} and \code{y}.
}
\description{
\code{euclidian} returns the lowest common denominator between two numbers.
}
\examples{
euclidian(1, 1)
euclidian(10, 1)
}
