\name{Lab 3-package}
\alias{Lab 3-package}
\alias{Lab 3}
\docType{package}
\title{
\packageTitle{Lab 3}
}
\description{
\packageDescription{Lab 3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab 3}
\packageIndices{Lab 3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Lab 3}

Maintainer: \packageMaintainer{Lab 3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
