% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3.R
\docType{package}
\name{lab3}
\alias{lab3}
\alias{lab3-package}
\title{lab3: A package to apply both the Euclidean and Dijkstra algorithms.}
\description{
The package provides 2 different functions:
euclidean(a,b) and dijkstra(g,i)
}
\section{lab3 functions}{

The function euclidean(a,b) finds the greatest common divisor of two numbers.
The function dijkstra(g,i) finds the shortest path from an initial node to every other node in a graph.
}

