\name{lab3package}
\alias{lab3package}
\docType{package}
\title{Lab 3 solution from ductr388 and wilwi856}
\description{
This package contains two functions used to solve lab 3 in the course 732A94.

The functions implemented are euclidean algorithm and dijkstra's algorithm 
}
\details{
Package: lab3package

Type: Package

Version: 1.0.0

Date: 2023-09-14

Licence: GPL (>=3)


}
\author{

Duc Tran Maintainer: <ductr388@student.liu.se> 

William Wiik Creator: <wilwi856@student.liu.se>
}
\references{
Information about the euclidean algorithm can be found 
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{here}.

Information about the dijkstra's algorithm can be found 
\href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{here}.
}



