% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{my_fun}
\alias{my_fun}
\title{Add together two numbers}
\usage{
my_fun()
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of `x` and `y`
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
}
