% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra.R, R/Dijkstra2.R, R/Euclidean.R
\docType{data}
\name{name}
\alias{name}
\title{Euclidean and Dijkstra Algorithms}
\format{An object of class \code{character} of length 1.}
\usage{
name

name

name
}
\arguments{
\item{graph}{(dataframe)}

\item{init_node}{(scalar)}

\item{graph}{(dataframe)}

\item{init_node}{(scalar)}

\item{x}{(numeric scalar)}

\item{y}{(numeric scalar)}
}
\value{
dist (vector)

dist (vector)

greatest common divisor (numeric scalar)
}
\description{
This package contains functions to compute the greatest 
    common divisor for two given numbers and find the shortest distance 
    from a source node to all the other nodes in a graph
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra2(wiki_graph3, "a")
euclidean(100, 1000)
}
\keyword{datasets}
