\name{trung440lab3-package}
\alias{trung440lab3-package}
\alias{trung440lab3}
\docType{package}
\title{
\packageTitle{trung440lab3}
}
\description{
\packageDescription{trung440lab3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{trung440lab3}
\packageIndices{trung440lab3}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{trung440lab3}

Maintainer: \packageMaintainer{trung440lab3}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
