\name{UgurcanFahaadLab3-package}
\alias{UgurcanFahaadLab3-package}
\alias{UgurcanFahaadLab3}
\docType{package}
\title{
\packageTitle{UgurcanFahaadLab3}
}
\description{
\packageDescription{UgurcanFahaadLab3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{UgurcanFahaadLab3}
\packageIndices{UgurcanFahaadLab3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{UgurcanFahaadLab3}

Maintainer: \packageMaintainer{UgurcanFahaadLab3}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
