% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.graph.R
\docType{data}
\name{wiki.graph}
\alias{wiki.graph}
\title{A dataset containing the wiki graph.}
\format{A data frame with 3 vectors:
\describe{
  \item{v1}{starting point of an edge}
  \item{v2}{end point of an edge}
  \item{w}{weight of the edge}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
wiki.graph
}
\description{
A dataset containing the wiki graph.
}
\keyword{datasets}

