% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Example Graph for Dijkstra's Algorithm}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{v1}{Starting node of an edge}
\item{v2}{Ending node of an edge}
\item{w}{Weight (distance) of the edge}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset containing the edges and weights of an example graph from Wikipedia.
}
\keyword{datasets}
