% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{wiki_graph2}
\alias{wiki_graph2}
\title{Wikipedia Graph2 (Character vertices)}
\format{The graph is a data frame with three variables (v1, v2 and w) 
  that contains the edges of the graph (from v1 to v2) with the weight 
  of the edge (w):
  
\describe{
  \item{v1}{vertex, starting point of edge in the graph}
  \item{v2}{vertex, ending point of edge in the graph}
  \item{w}{weight of the edge}
}}
\usage{
wiki_graph2
}
\description{
A sample graph with six vertices and their weighted edges
}
\keyword{datasets}
