% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algorithm for the shortest path.}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{It must be data frame. The graph is a data frame including three variables v1,v2,w and contains the edges of the graph from v1 to v2 and the weight of the edge w.}

\item{init_node}{It must be numeric scalar. The initial node that the algorithm will start calculating the shortest path.}
}
\value{
It will be a vector with the shortest paths from the initial node.
}
\description{
Dijkstra Algorithm for the shortest path.
}
