% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{dijkstra}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame with 3 varibles.}

\item{init_node}{A numeric value.}
}
\value{
The shortest path from inital node to every other node in the graph.
}
\description{
dijkstra
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
