% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Add together two numbers}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of `x` and `y`
}
\description{
Add together two numbers
}
\examples{
print("hi")

}
