% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra}
\source{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame for ...}

\item{init_node}{The init_node node is}
}
\value{
Returns a vector with every shortest path from
}
\description{
Sample text
}
\examples{
dijkstra(wiki_graph, 1) ## c(0, 7, 9, 20, 20, 11)
}
