% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3_Aset_Arnab.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Must be a \code{data.frame} object with three variables (\code{v1}, \code{v2} and \code{w}) that
contains the edges of the \code{graph} (from \code{v1} to \code{v2}) with the weight of the edge (\code{w}).}

\item{init_node}{'Initial node' must be a numeric scalar that exists in the graph description}
}
\value{
Vector with the distances from \code{init_node} to every other node in the graph
}
\description{
Algorithm takes \code{graph} and \code{init_node} and calculates the shortest path from the initial node to every other node in the graph
}
\examples{
wiki_graph <- data.frame(
  v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
  v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
  w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))

dijkstra(wiki_graph, 1)
}
\references{
Dijkstra's algorithm \url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
