% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A dataset representing the graph in which the algorithm should search for the shortest paths}

\item{init_node}{The starting node from which the shortest paths are calculated}
}
\value{
Distances of the shortest paths between the given starting node to each other node
}
\description{
The algorithm finds the distance  of the shortest path between a given node to every other node in the graph
}
\references{
https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm
}
