% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's shortest path algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame with three variables (v1, v2, w) that contains the edges
of the graph (from v1 to v2) with the weight of the edge (w).}

\item{init_node}{A numeric scalar that exists in the graph.}
}
\value{
A vector containing the shortest path distance from \code{init_node}
        to every other node in the graph.
}
\description{
This function implements Dijkstra's algorithm, which is a graph search algorithm
that solves the single-source shortest path problem for a graph with non-negative
edge weights, producing a shortest path tree.
}
\examples{
wiki_graph <- data.frame(
  v1 = c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
  v2 = c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
  w = c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)
)
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)

}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
