% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{dijkstra's Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A list representing the adjacency list of the graph.}

\item{init_node}{The starting node for the algorithm.}
}
\value{
A list containing the shortest paths from the source to all other nodes.
}
\description{
This function implements Dijkstra's algorithm for finding the shortest paths from a source node to all other nodes in a graph.
}
\examples{
graph <- list(
  A = list(B = 1, C = 4),
  B = list(A = 1, C = 2, D = 5),
  C = list(A = 4, B = 2, D = 1),
  D = list(B = 5, C = 1)
)
dijkstra(graph, "A")

}
\references{
Wikipedia: \href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{Dijkstra's algorithm}
}
