% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dijkstra}
\alias{dijkstra}
\title{dijkstra}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{data_frame with three colums}

\item{init_node}{node from whitch we want to calculate the distancies}
}
\description{
This function implements dijkstra
}
