% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Calculate shortest path distances using Dijkstra's algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A \code{\link[base]{data.frame}} representing the graph.
Each row represent an edge and the data fram must contain the following
columns:
\describe{
\item{\code{v1}}{A \code{vector} containing start nodes}
\item{\code{v2}}{A \code{vector} containing destination nodes}
\item{\code{w}}{A \code{vector} containing the weights from v1_i to v2_i}
}
All edge weights must be non-negative.}

\item{init_node}{An integer specifying the initial node from which
shortest path distances are calculated. The node must occur in either
\code{v1} or \code{v2}.}
}
\value{
A numeric \code{link[base]{vector}} containing the shortest
path distance from \code{init_node} to all other nodes.
}
\description{
Dijkstra's algorithm finds the shortest path distances from one
initial node to all other nodes in a directed weighted graph.
}
\examples{
data(wiki_graph)
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 6)
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
