% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Computes the shortest path from a given node to every other node in a graph with only positive edges}
\usage{
dijkstra(graph, start_node)
}
\arguments{
\item{graph}{a graph with only positively weighted edges}

\item{start_node}{the node which the shortest path to all the other nodes is computed from}
}
\value{
It returns a vector with the lengths of the shortest path to all nodes in the graph from start_node. The index indicates which node the path is to
}
\description{
Computes the shortest path from a given node to every other node in a graph with only positive edges
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
