% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{dijkstra}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{a}{dataframe and a starting node (scalar)}
}
\value{
a vector of shortest distances to each other node.
}
\description{
takes a graph and starting node.
Runs dijkstra's algorithm,
i.e. finds the shortest distance to each other node.
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}

