% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\alias{dij}
\title{Dijkstra Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A dataframe containing the nodes and weights of each vertex in the graph.}

\item{init_node}{A number to identify the inital node.}
}
\value{
Vector of distances between \code{init_node} and every other node in \code{graph}
}
\description{
\code{dijkstra} finds the shortest path from a given node to every other node in a given graph.

It picks the unvisited vertex with the lowest distance, calculates the distance through it to each unvisited neighbor, and updates the neighbor's distance if smaller.
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
