% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{This function takes a graph and an initial node and calculates the shortest path from the initial node to every other node in the graph.}
\source{

}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{R data frame.}

\item{init_node}{A number.}
}
\value{
The shortest path from the `init_node` to every other node in the `graph`.
}
\description{
This function takes a graph and an initial node and calculates the shortest path from the initial node to every other node in the graph.
}
\examples{
dijkstra(graph, 1)
dijkstra(graph, 3)
}
