% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{1.1.2 Dijkstra algorithm implementation}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{= This takes a data frame as a parameter which includes three variable named v1, v2, w.
Here v1 and v2 contains edges and w contains weighted distance from v1 to v2 edges.}

\item{init_node}{= It takes a numeric integer value from where we will calculate the shortest distance to every other nodes in our given graph.}
}
\value{
The function returns a vector with the shortest distances from init_node.
}
\description{
This function calculates shortest distance from an initial node to all other nodes in a given data frame.
}
\examples{
wiki_graph <-  data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)

}
\references{
\url{en.wikipedia.org/wiki/Dijkstra_algorithm}
}
