% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algorithm modified}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Dataframe}

\item{init_node}{Numeric scalar}
}
\value{
Numeric vector representing the shortest distance form 
        init_node to all the other nodes in graph
}
\description{
This function finds the shortest distance from a source 
             node to all the other nodes in a graph
}
\details{
Advanced Programming in R - Lab 3
Dijkstra algorithm
}
\examples{
dijkstra(wiki_graph, 1)
}
