% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{Dijkstra}
\alias{Dijkstra}
\alias{dijkstra}
\title{Dijkstras algorithm}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame containing three variables v1, v2, and w that contains
the edges of the graph (from v1 to v2) with the weight of the edge (w).}

\item{init_node}{A numeric scalar that exists in the graphs (alas in the v1, v2).}
}
\value{
A vector consisting of the closest path to each graph from the initial node
}
\description{
Find the shortest path between nodes in a weighted graph.
The algorithm itself and pseudo-code is from
https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm.
}
