% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Computes the least path costs to all nodes in a directed graph from the initial node.}
\format{graph
\describe{
   \item{v1}{Vector of nodes for edges start positions.}
   \item{v2}{Vector of nodes for edges end positions.}
   \item{w}{Vector of edge weights.}
}}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}
}
\usage{
dijkstra(graph, node_init)
}
\arguments{
\item{graph}{A directed graph which is supposed to compute the least path of different nodes.}

\item{node_init}{A node which is chosen for computing the shortest path.}
}
\value{
Vector of path costs to every node from \code{node_init}.
}
\description{
Computes the least path costs to all nodes in a directed graph from the initial node.It assumes the nodes are of type numeric from 1 to |V| (number of nodes).
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)

}

