% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3.r
\name{dijkstra}
\alias{dijkstra}
\title{A simple implementation of the Dijkstra algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{a data.frame with three columns v1,v2,w}

\item{init_node}{the initial node, an integer}
}
\value{
a vector of distances
}
\description{
A simple implementation of the Dijkstra algorithm
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
