% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's Algorithm for Shortest Paths}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame with three columns:
\describe{
  \item{v1}{Start node of an edge (numeric).}
  \item{v2}{End node of an edge (numeric).}
  \item{w}{Weight of the edge (numeric).}
}}

\item{init_node}{A numeric scalar giving the starting node.}
}
\value{
A numeric vector of shortest path distances from `init_node` to every node.
}
\description{
Calculates the shortest path distances from a starting node to all other
nodes in a directed graph.
}
\examples{
wiki_graph <- data.frame(
  v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
  v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
  w =c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9)
)
dijkstra(wiki_graph, 1) # [1] 0 7 9 20 20 11

}
\references{
<https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm>
}
