% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{Graph}{A data Frame}

\item{Init_node}{A numeric scalar}
}
\value{
Numeric Vector with the shortest distance between nodes
}
\description{
Finding the shortest distance between nodes using the Dijkstra algorithem
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
