% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkastra Algorithm implementation}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph, initialNode}{a dataframe containing columns v1,v2,w and a starting node}
}
\value{
The vector containing shortest path from starting node to other nodes in a graph
}
\description{
Find shortest path of a given node to other nodes in a graph
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
