% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstras algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{Data frame consisting of starting locations, end locations and weights.}

\item{init_node}{Initial starting node.}
}
\value{
Returns a vector with shortest distance from starting node to all other nodes.
}
\description{
Calculates the shortest distance from initial node to all other nodes using dijkstra's algorithm.
}
\details{
Further information on the dijkstras algorithm can be found at 
\href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{Wikipedia}
}
