% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Djikstra Algorithm}
\usage{
dijkstra(data, initial_node)
}
\arguments{
\item{data}{A data.frame with three variables (v1, v2 and w) that contains the nodes of the graph (from v1 to v2) with the length of the node (w)}

\item{initial_node}{The node to start with.}
}
\value{
The shortest path from the initial node to every other node.
}
\description{
Takes a graph and an initial node, then calculates the shortest path from the initial node to every other node in the graph.
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
