% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{dijkstra}
\alias{dijkstra}
\title{find the shortest path between the initial node and other node}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{data.frame}

\item{init_node}{A number}
}
\value{
a vector containing these shortest path
}
\description{
find the shortest path between the initial node and other node
}
