% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric value.}

\item{b}{A numeric value.}
}
\value{
The greatest common divisor of a and b.
}
\description{
The best experience with the lab 3

Euclidean algorithm
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
