% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\source{
Euclidean algorithm. (September, 2025). In Wikipedia. Retrieved from https://en.wikipedia.org/w/index.php?title=Euclidean_algorithm&oldid=1304987228
}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{a whole number of length 1}

\item{b}{a whole number of length 1}
}
\value{
The GCD
}
\description{
A short description...

The Euclidean algorithm finds the greatest common divisor, GCD between to whole numbers, a and b.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)

}
