% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3_Aset_Arnab.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A positive numeric scalar.}

\item{b}{A positive numeric scalar.}
}
\value{
A single numeric value representing the greatest common divisor of \code{a} and \code{b}.
}
\description{
Calculates the greatest common divisor (GCD) of two positive integers
using the Euclidean algorithm.
}
\examples{
euclidean(100, 1000)
euclidean(123612, 13892347912)
}
\references{
Euclidean algorithm \url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
