% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{The first number}

\item{b}{The second number}
}
\value{
Greatest common divisor of the two input numbers
}
\description{
The algorithm finds the greatest common divisor of two given numbers
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
