% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find greatest common divisor between two numbers}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A natural number.}

\item{b}{A natural number.}
}
\value{
The largest natural number that divides both a and b.
}
\description{
It calculates the greatest common divisor of \item{a} and \item{b} using the Euclidean algorithm. 
}
\examples{
euclidean(1000, 100)
euclidean(10839893, 13238)
}
\references{

Wikipedia (2017). \emph{Eucliedean algorithm - Wikipedia, The Free Encyclopedia.}
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm} , accessed 2017-09-14.

}