% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find greatest common divisor between two numbers}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
Greatest common division using Euclidean's algorithm
}
\description{
Find greatest common divisor between two numbers
}
\examples{
euclidean(1000, 100)
euclidean(10839893, 13238)
}
