% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm}
\usage{
euclidean(num1, num2)
}
\arguments{
\item{num1}{An integer representing the first number.}

\item{num2}{An integer representing the second number.
Both numbers are used to calculate their greatest common divisor (GCD).}
}
\value{
The greatest common divisor.
}
\description{
This function implements  is an efficient method for computing the greatest common divisor (GCD) of two integers (numbers).
}
\examples{
euclidean(num1 = 100, num2 = 1000)

}
\references{
Wikipedia: \href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{Euclidean Algorithm}
}
