% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanwithwhile.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{Numeric scalar or integer}

\item{b}{Numeric scalar or integer}
}
\value{
Greatest common divisor of a and b
}
\description{
Euclidean Algorithm
}
\examples{
 euclidean(123612, 13892347912) 
           euclidean(100, 1000)

}
