% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{The Euclidian algorithm can find the greatest common divisor of two numbers.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{Integer number}

\item{b}{Interger number}
}
\value{
An interger represent the greatest common divisor of two numbers.
}
\description{
The Euclidian algorithm can find the greatest common divisor of two numbers.
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
