% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{euclidean}
\alias{euclidean}
\title{euclidean}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{integer scalar}

\item{b}{integer scalar}
}
\value{
integer scalar
}
\description{
This function implements euclidean algorithm to calculate the MCD of two given numbers (a, b).
This function iteratively calculates the remainder of the whole division between the previous two remainders.
The output is the last non-zero remainder
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
