% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm (Greatest Common Divisor)}
\usage{
euclidean(prev, remainder)
}
\arguments{
\item{prev}{The first numeric integer input (usually the larger number).}

\item{remainder}{The second numeric integer input (usually the smaller number).}
}
\value{
A single numeric integer representing the greatest common divisor
of the two provided inputs.
}
\description{
This function computes the greatest common divisor (GCD) of two integers
using the Euclidean algorithm. The algorithm repeatedly replaces the
remainder until it becomes zero. The last non-zero remainder is the GCD.
}
\examples{
euclidean(462, 1071)

}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
