% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm
Computes the greatest common divisor of two integers.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric scalar}

\item{b}{A numeric scalar.}
}
\value{
A numeric scalar: the greatest common divisor of \code{a} and
\code{b}
}
\description{
Euclidean algorithm
Computes the greatest common divisor of two integers.
}
\details{
The algorithm repeatedly replaces the pair (a,b) with (b,r), where r is the
remainder of a divided by b, until b is zero. The remaining value of a is
then the greatest common divisor. the sing is removed at the end, the result
is always non-negative
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)

}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
