% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.r
\name{euclidean}
\alias{euclidean}
\title{Calculate the greatest common divisor using the Euclidean algorithm}
\usage{
euclidean(M, N)
}
\arguments{
\item{M}{A numeric value of length one containing the first number.}

\item{N}{A numeric value of length one containing the second number.}
}
\value{
The greatest common divisor of M and N.
}
\description{
The Euclidean algorithm calculates the greatest common divisor
of two numbers by repeatedly replacing the larger number with the
remainder obtained when dividing the larger number by the smaller one.
This process continues until the remainder is zero. The last non-zero
remainder is the greatest common divisor.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
