% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\alias{eucl}
\alias{euclidean_algorithm}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
Greatest common divisor between \code{a} and \code{b}
}
\description{
\code{euclidean} finds the greatest common divisor of two integers.
}
\examples{
euclidean(123612,13892347912)
euclidean(100,1000)
}
\seealso{
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{Wikipedia}
}
\concept{\code{\link{dijkstra}}}
