% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(num1, num2)
}
\arguments{
\item{num1}{The first number to find the greatest common divisor from}

\item{num2}{The first number to find the greatest common divisor from}
}
\value{
returns the greatest common divisor of the two input numbers
}
\description{
The algorithm finds the greatest common divisor of two given numbers
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
