% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find the greatest common divisor of two numbers.}
\source{
\url{http://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
Euclidean distance of `a` and `b`
}
\description{
Find the greatest common divisor of two numbers.
}
\examples{
euclidean(12, 12)
euclidean(100, 1000)
}
