% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean.R
\name{euclidean}
\alias{euclidean}
\title{1.1.1 euclidean function}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{is the first input of function. Any integer is accepted.}

\item{b}{is the second input of function. Any integer is accepted.}
}
\value{
The greatest common divisor of two integers using Euclidean division
}
\description{
The function is simple, we just needed to define a method to assert before
calculations Always pay attention to main conditions defined by function that can
stops your code. This function automatically swap inputs if the second parameter is lower than the first
}
\examples{
temp1=euclidean(100L,1000L)
temp2=euclidean(123612,13892347912)
temp3=euclidean(100,1000)
print(c(temp1,temp2,temp3))
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
