% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\source{
\url{"https://en.wikipedia.org/wiki/Euclidean_algorithm"}
}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{Numeric scalar}

\item{y}{Numeric scalar}
}
\value{
Numeric scalar which is greatest common divisor of x and y
}
\description{
This function computes the greatest common divisor for 
             any two given numbers
}
\details{
Advanced Programming in R - Lab 3
Euclidean algorithm
}
\examples{
euclidean(100, 1000)
}
