% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{Euclidean}
\alias{Euclidean}
\alias{euclidean}
\title{Euclidean algorithm}
\source{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{numeric scalar or integer}

\item{y}{numeric scalar or integer}
}
\value{
The Greatest common divisor of x and y
}
\description{
Euclidean algorithm which finds the greatest common divisor between two numeric
scalars or integers. https://en.wikipedia.org/wiki/Euclidean_algorithm
}
