% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\source{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\usage{
euclidean(a, b)
}
\arguments{
\item{\code{a}}{A positive integer.}

\item{\code{b}}{A positive integer.}
}
\value{
The greatest common divisor of \code{a} and \code{b}.
}
\description{
Finds the greatest common divisor of two integers by using the Euclidean algorithm.
}
\examples{
euclidean(100, 1000)
euclidean(1000, 100)

}

