% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3.r
\name{euclidean}
\alias{euclidean}
\title{Algorithm for finding the greatest common divisor, Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{integer}

\item{b}{integer}
}
\value{
an integer which is the gcd(a,b)
}
\description{
Algorithm for finding the greatest common divisor, Euclidean algorithm
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
