% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm for Greatest Common Divisor}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric scalar or integer.}

\item{b}{A numeric scalar or integer.}
}
\value{
An integer giving the greatest common divisor of `a` and `b`.
}
\description{
Computes the greatest common divisor (GCD) of two numbers
using the Euclidean algorithm.
}
\examples{
euclidean(123612, 13892347912) # [1] 4
euclidean(100, 1000) # [1] 100

}
\references{
<https://en.wikipedia.org/wiki/Euclidean_algorithm>
}
