% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Eculidean function
it compute the GCD of two numeric}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{is a numeric}

\item{b}{is a numeric}
}
\value{
GCD
}
\description{
Eculidean function
it compute the GCD of two numeric
}
\examples{
euclidean (100, 1000)

}
