% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3.R
\name{euclidean}
\alias{euclidean}
\title{euclidean Algorithm}
\usage{
euclidean(m, n)
}
\arguments{
\item{a}{number}

\item{b}{number}
}
\value{
greatest common divisor of two numbers
}
\description{
euclidean algorithm to find the greatest common divisor of two numbers
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
