% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean method to find greatest Common divisor}
\usage{
euclidean(Num1, Num2)
}
\arguments{
\item{Num1, Num2}{Two numbers for which greatest common divisor is to be calculated}
}
\value{
The greatest common divisor of `Num1` and `Num2`.
}
\description{
it takes two numbers and calculate the greatest common divisor of given 2 numbers
}
\examples{
euclidean(1, 1)
euclidean(10, 1)
euclidean(100, -1000)
euclidean(-100, 1000)
euclidean(-10, -100)
}
