% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidian algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A scalar.}

\item{b}{A scalar.}
}
\value{
Returns a scalar of the greatest common divisor.
}
\description{
Calculates the greatest common divisor if two numbers.
}
\details{
Further information on the euclidian algorithm can be found at Wikipedia 
(https://en.wikipedia.org/wiki/Euclidean_algorithm).
}
