% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidian algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A integer.}

\item{b}{A integer.}
}
\value{
Returns a integer of the greatest common divisor.
}
\description{
Calculates the greatest common divisor between two integers.
}
\details{
Further information on the euclidian algorithm can be found at 
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{Wikipedia}
}
