% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{find the greatest common divisor of two numbers.
https://en.wikipedia.org/wiki/Euclidean_algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The greatest common divisor of \code{a} and \code{b}.
}
\description{
find the greatest common divisor of two numbers.
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
