% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{find the greatest common divisor of two numbers}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
the greatest common divisor of two numbers
}
\description{
find the greatest common divisor of two numbers
}
