% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/euclidian.R
\name{euclidian}
\alias{euclidian}
\title{euclidian}
\usage{
euclidian(a, b)
}
\arguments{
\item{a:=scalar,}{b:=scalar}
}
\value{
GCD of a and b
}
\description{
Finds GCD's of two arbitrary scalars.
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}

