\name{lab3-package}
\alias{lab3-package}
\alias{lab3}
\docType{package}
\title{
\packageTitle{lab3}
}
\description{
\packageDescription{lab3}
}
\details{
Package contains dijkstra and euclidian algorithms
}
\author{
%% ~~ Ana Stanojcic & 
\packageAuthor{Ana Stanojcic, Sofia Danielson}

Maintainer: \packageMaintainer{Ana Stanojcic, Sofia Danielson}
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
https://sv.wikipedia.org/wiki/Dijkstras_algoritm
}
\keyword{Euclidian, Dijkstra}

%% Uncomment below to imitate parts of library(help = lab3)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{lab3}}
%\section{Documentation Index}{\packageIndices{lab3}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
