\name{lab3-package}
\alias{lab3-package}
\alias{lab3}
\docType{package}
\title{
\packageTitle{lab3}
}
\description{
\packageDescription{lab3}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{lab3}

Maintainer: \packageMaintainer{lab3}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = lab3)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{lab3}}
%\section{Documentation Index}{\packageIndices{lab3}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
