\name{Lab3-package}
\alias{Lab3-package}
\alias{Lab3}
\docType{package}
\title{
Distance computation
}
\description{
Computes the euclidean distance and dijkstra algorithm
}
\details{
\tabular{ll}{
Package: \tab Lab3\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-07\cr
License: \tab What license is it under?\cr
}
euclidean(x,y)
dijkstra(graph, init_node)
}
\author{
Martina Sandberg, Caroline Svahn

Maintainer: Who to complain to <carsv733@student.liu.se>

}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}

\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
