% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3-package.R
\docType{package}
\name{lab3-package}
\alias{lab3}
\alias{lab3-package}
\title{lab3: Practice for making packages in R}
\description{
This package has two functions: finding euclidean distance and Djikstra's algo implementation
}
\author{
\strong{Maintainer}: Usama Nadeem \email{usamanadeem321@gmail.com}

Authors:
\itemize{
  \item Muhammad Umair \email{Umair.raja.1422@gmail.com}
}

}
\keyword{internal}
