\name{lab3-package}
\alias{lab3-package}
\alias{lab3}
\docType{package}
\title{
\packageTitle{lab3}
}
\description{
\packageDescription{lab3}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab3}
\packageIndices{lab3}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lab3}

Maintainer: \packageMaintainer{lab3}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
