% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3.R
\docType{package}
\name{lab3}
\alias{lab3}
\alias{lab3-package}
\title{lab3}
\description{
A package to apply both the Euclidean and Dijkstra algorithms.
}
\details{
The package provides 2 different functions:
euclidean(a,b) and dijkstra(g,i)
}
\section{Functions}{

The function euclidean(a,b) finds the greatest common divisor between a and b.

The function dijkstra(graph,initial_node) finds the shortest path from an initial_node to every other node in graph.
}

