% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-lab3.R
\docType{package}
\name{lab3}
\alias{lab3}
\title{A package that implements the Euclidean and Dijkstra algorithm}
\description{
This package includes a function executing the Euclidean algorithm for finding the greatest common divisor of two input numbers.
It furthermore includes a function executing the Dijkstra algorithm. The algorithm provides the distance of the shortest path in an input graph from a starting node that is defined as an input to the function.
}
\section{euclidean function and dijkstra function}{
NA
}

