% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\docType{package}
\name{Lab3}
\alias{Lab3}
\alias{Lab3-package}
\title{Implementation of Euclidean and Dijkstra algorithm}
\description{
Implementation of Euclidean algorithm: finds the greatest common
divisor of two numbers. 

Implementation of Dijkstra algorithm: finds the shortest
paths within a directed graph.
}
\author{
Rasmus Holm

Zhendong Wang

Karolina Ziomek
}

