% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3.R
\docType{package}
\name{lab3}
\alias{lab3}
\alias{lab3-package}
\title{Euclidean package.}
\description{
The Euclidean package provides one functions:
euclidean.
}
\section{Euclidean functions}{

 it returens the GCD
}

