% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab3.R
\docType{package}
\name{lab3}
\alias{lab3}
\title{lab3: A package for Euclidian algorithm and dijkstra}
\description{
The lab3 package provides two functions: euclidian and dijkstra.
}
\section{euclidian}{
 The euclidian function find the greatest common divisor
  of two numbers.
}

\section{dijkstra}{
 The dijkstra function takes a graph and an initial node
  and calculates the shortest path from the initial node to every other node
  in the graph.
}

