\name{lab3ab-package}
\alias{lab3ab-package}
\alias{lab3ab}
\docType{package}
\title{
\packageTitle{lab3ab}
}
\description{
\packageDescription{lab3ab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab3ab}
\packageIndices{lab3ab}

}
\author{
\packageAuthor{lab3ab}

Maintainer: \packageMaintainer{lab3ab}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{

}
\examples{
}
