\name{lab3ab-package}
\alias{lab3ab-package}
\alias{lab3ab}
\docType{package}
\title{
\packageTitle{lab3ab}
}
\description{

\packageDescription{lab3ab} 

Includes two algorithms (Euclidian and Dijkstra) and a dataset with three variables that contains the edges of the graph with the weight of the edge. 

}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab3ab}

\packageIndices{lab3ab}

}

\author{

\packageAuthor{lab3ab}

Maintainer: \packageMaintainer{lab3ab}

}

\references{

Wikipedia (2017). \emph{Eucliedean algorithm - Wikipedia, The Free Encyclopedia.}
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm} , accessed 2017-09-14.

Wikipedia (2017). \emph{Dijkstra's algorithm - Wikipedia, The Free Encyclopedia.}
\url{https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}, accessed 2017-09-14.

}

\keyword{ package }

