\name{lab3package-package}
\alias{lab3package-package}
\alias{lab3package}
\docType{package}
\title{
\packageTitle{lab3package}
}
\description{
\packageDescription{lab3package}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{lab3package}
\packageIndices{lab3package}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{lab3package}

Maintainer: \packageMaintainer{lab3package}
}
\references{
Reference wikipedia pages
}
\keyword{ package }
\seealso{
Removed see also
}
\examples{
print("Example text")
}


