% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra.R, R/Euclidean.R
\docType{data}
\name{name}
\alias{name}
\title{Dijkstra algorithm}
\format{An object of class \code{character} of length 1.}
\usage{
name

name
}
\arguments{
\item{graph}{(dataframe)}

\item{init_node}{(scalar)}

\item{x}{(numeric scalar)}

\item{y}{(numeric scalar)}
}
\value{
dist (vector)

greatest common divisor (numeric scalar)
}
\description{
This package computes the shortest distance from a source node to all the other nodes in a graph

This package computes the greatest common divisor for two given numbers
}
\examples{
dijkstra(wiki_graph, 1)
euclidean(100, 1000)
}
\keyword{datasets}
