% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_graph 
A data frame contains graph data with three variables.}
\format{
A data frame with v1, v2, w variables and 18 rows of data.
\describe{
\item{v1}{Nodes of the graph}
\item{v2}{Nodes of the graph}
\item{w}{Weight of the nodes, distance}

}
}
\source{
https://en.wikipedia.org/wiki/Dijkstra%27s algorithm
}
\usage{
wiki_graph
}
\description{
Wiki_graph 
A data frame contains graph data with three variables.
}
\keyword{datasets}
