% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wikipedia example graph}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{v1}{Numeric. The starting node of the edge.}
\item{v2}{Numeric. The ending node of the edge.}
\item{w}{Numeric. The weight (cost) of the edge between v1 and v2.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset representing a small graph, used to demonstrate graph
algorithms such as Dijkstra's algorithm.
}
\keyword{datasets}
