% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\format{
A data frame with 3 columns, v1, v2 and w.
\describe{
Each row of the dataframe represents and arc in the graph. The arc extends from the node v1 to the node v2 and its corresponding weight is w.

}
}
\source{
Dijkstra’s algorithm. (September, 2025). Retrieved from \href{https://en.wikipedia.org/w/index.php?title=Dijkstra\%27s_algorithm&oldid=1306611316}{Wikipedia}
}
\usage{
data(wiki_graph)
}
\description{
The weighted graph from the example on the Dijkstra Wikipedia article.
}
\keyword{datasets}
