% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_graph}
\format{
A data frame with 3 columns, v1, v2 and w.
\describe{
Each row of the dataframe represents and arc in the graph. From v1 to v2 and its corresponding weight w.

}
}
\source{
https://en.wikipedia.org/wiki/Dijkstra's_algorithm
}
\usage{
wiki_graph
}
\description{
The weighted graph from the dijkstra wikipedia page.
}
\keyword{datasets}
