% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3_Aset_Arnab.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Example Graph Dataset for Dijkstra's Algorithm}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{Numeric: start node of the edge}
  \item{v2}{Numeric: end node of the edge}
  \item{w}{Numeric: weight of the edge}
}
}
\usage{
wiki_graph
}
\description{
This dataset represents a weighted undirected graph with 6 nodes and edges with weights.
It can be used to test the Dijkstra function in this package.
}
\references{
Wikipedia page for graphs: \url{https://en.wikipedia.org/wiki/Graph_(discrete_mathematics)}
}
\keyword{datasets}
