% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki Graph}
\format{
A dataframe with 3 columns and 18 rows:
\describe{
\item{v1}{an edge starting from vertex / node 1}
\item{v2}{to the target vertex / node 2}
\item{w}{weight of the edge between v1 and v2}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset containing the nodes and edges of the graph from the first example on the Wikipedia page of Dijkstra algorithm.
}
\keyword{datasets}
