% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Weighted graph with 6 vertices}
\format{A data frame with 3 vector as columns :
\describe{
  \item{v1}{Starting vertices of each edges}
  \item{v2}{Ending vertices of these edges.}
  \item{w}{Weights of each edge}
  
}}
\usage{
wiki_graph
}
\description{
Weighted graph with 6 vertices
}
\keyword{datasets}
