% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\format{
A data frame with 3 columns and N rows.
}
\usage{
data(wiki_graph)
}
\description{
This dataset contains information about wiki_graph. Vertices v1, v2, and weight w
}
\details{
\describe{
  \item{v1}{An integer vector representing the from vertex.}
  \item{v2}{An integer vector representing the to vertex.}
  \item{w}{An integer vector the weight/cost of the transition from v1 to v2.}
}
\href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{this website}
}
\keyword{datasets}
