% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Distances w between 6 nodes v.}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{v1}{point 1}
\item{v2}{point 2}
\item{w}{weight or distance from point 1 to point 2}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra's algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset containing the pathway showed on top of the wikipedia page from Dijkstra's algorithm
}
\keyword{datasets}
