% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki Graph Example Dataset}
\format{
A data.frame with 3 columns.
}
\source{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\usage{
wiki_graph
}
\description{
A dataset representing a weighted directed graph. It contains three variables:
\itemize{
  \item \code{v1} Starting node of an edge
  \item \code{v2} Ending node of an edge
  \item \code{w}  Weight of the edge between the nodes
}
}
\details{
The dataset is commonly used to demonstrate Dijkstra's algorithm for finding
the shortest paths in a graph.
}
\examples{
data(wiki_graph)
head(wiki_graph)
}
\keyword{datasets}
