% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{A graph dataset with 6 vertecies and 9 edges.}
\format{
\subsection{\code{wiki_graph}}{

A data frame with 18 rows and 3 columns:
\describe{
\item{v1}{Vertex 1}
\item{v2}{Vertex 2}
\item{w}{weight}
}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A graph dataset with 6 vertecies and 9 edges.
}
\keyword{datasets}
