% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{required wiki graph data}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{An integer vector containing representations for nodes.}
  \item{v2}{An integer vector containing representation for nodes that are connected to the nodes in v1 via an edge.}
  \item{w}{An integer vector containing the weights for the edges between the nodes in v1 and v2.}
}
}
\source{
Generated for demonstration purposes.
}
\usage{
wiki_graph
}
\description{
This graph is an undirected graph with 6 nodes connected with posetively weighted edges
}
\examples{
# Load the data
data(wiki_graph)
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\keyword{datasets}
