% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wiki_graph.R
\name{wiki_graph}
\alias{wiki_graph}
\title{Example dataset to the dijkstra algorithm.}
\format{a dataframe with 3 columns. v1 corresponding to "from-nodes"
. v2 corresponding of "to-nodes" and w the distance between each such.}
\description{
a dataframe of connections between nodes
}
\references{
www.wikipedia.com
}

