% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_data}
\format{
A data frame with variables
\describe{
  \item{v1}{first}
  \item{v2}{second}
  \item{w}{weight}
  }
}
\usage{
wiki_graph
}
\description{
Contains good data
}
\examples{
data(wiki_data)
}
\references{
{wikipedia, skilnad from source}
}
\keyword{datasets}
