% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wikipedia Graph}
\format{\code{wiki_graph} is data.frame with 18 rows and 3 variables:
\describe{
\item{v1}{starting node, no unit}
\item{v2}{ending node, no unit}
\item{w}{weight of the vertex between \code{v1} and \code{v2} , no unit}
}}
\usage{
wiki_graph
}
\description{
A dataset containing the graph used as example in Wikipedia.
}
\keyword{datasets}
