% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Nodes and weights}
\format{A data frame with 18 rows and 3 variables}
\usage{
data(wiki_graph)
}
\description{
A dataset containing nodes and wieghts for a graph.
The variables are as follows:
}
\details{
\itemize{
  \item v1 vector with nodes
  \item v2 vector with nodes
  \item w vector with weights
}
}
\keyword{datasets}

