% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_graph data frame as a dataset}
\description{
Wiki_graph data frame as a dataset
}
\references{
\url{https://en.wikipedia.org/wiki/File:Wiki_graph.jpg}
}
\author{
Adeel \email{adeah090@student.liu.se}
}
\keyword{data}
