% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\format{
A data frame with 3 variables:
\describe{
  \item{v1}{First node}
  \item{v2}{Second node}
  \item{w}{Distance from first node to second one}
  ...
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset containing the distances from one node to other with their weighted values.
}
\keyword{datasets}
