% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{A graph with 6 nodes.}
\format{A data frame with 18 rows and 3 columns:
 \describe{
  \item{v1}{numer of a node}
  \item{v2}{number of the related node to
  corresponding element in v1}
  \item{w}{weight of the edge}
  }}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
This dataset contains details about a graph structure. This graph has 6 nodes
whith different weight in each edge to another nodes.
}
\keyword{datasets}
