% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Graph data from wikipedia}
\format{
\subsection{\code{wiki_graph}}{

A data frame with 18 rows and 6 columns:
\describe{
\item{v1}{nodes, number of each node depends on number of neighbours of the node}
\item{v2}{connection between the nodes in v1}
\item{w}{distance between each pair of node}
...
}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
Dijkstra algorithm example from wikipedia
}
\keyword{datasets}
