% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{A data set containing the example directed graph which is shown in wiki website.}
\format{A data frame with 3 vectors:
\describe{
  \item{v1}{Start nodes of the edges.}
  \item{v2}{End nodes of the edges.}
  \item{w}{Edge weights.}
}}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra's_algorithm}
}
\usage{
wiki_graph
}
\description{
A data set containing the example directed graph which is shown in wiki website.
}
\keyword{datasets}

