% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Example graph for Dijkstra's algorithm}
\format{
data.frame
\describe{
  \item{v1}{Start node of the directed edge (numeric).}
  \item{v2}{End node of the directed edge (numeric).}
  \item{w}{Weight (numeric).}
}
}
\usage{
wiki_graph
}
\description{
A small directed, weighted graph used in examples of Dijkstra’s algorithm.
}
\examples{
data(wiki_graph)
head(wiki_graph)
}
\references{
<https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm>
}
\keyword{datasets}
