% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki-graph}
\format{A data frame}
\usage{
data(wiki_graph)
}
\description{
A dataset containing three variables.
}
\references{
\href{http://www.r-tutor.com/r-introduction/data-frame}{R tutorial}
}
\keyword{dataset}
