% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{A wikipedia graph dataset}
\format{
18 objects, 3 variables
\describe{
  \item{v1}{edge of graph}
  \item{v2}{edge of graph}
  \item{w}{weight of edge}
  ...
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra}
}
\usage{
wiki_graph
}
\description{
A dataset containing the plot for a graph from wikipedia page
diamonds.
}
\keyword{datasets}
