% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Example data for dijkstra's function}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{starting node}
  \item{v2}{end node}
  \item{w}{weight between starting and end node}
}
}
\usage{
wiki_graph
}
\description{
A dataset containing the weight between two nodes
}
\details{
Data set is used for dijkstra's function, more information about dijkstras algorithm 
can be found at: 
\href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{Wikipedia}
}
\keyword{datasets}
