% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki Graph}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{source node}
  \item{v2}{destination node}
  \item{w}{weight or length}
}
}
\usage{
wiki_graph
}
\description{
A Graph(data.frame) with three variables (v1, v2 and w) that 
contains the nodes of the graph (from v1 to v2) with the length of the nodes (w)
}
\keyword{datasets}
