% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{example graph}
\format{
## 'wiki_graph'
A data frame with 3 rows and 18 columns:
}
\source{
<https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm>
}
\usage{
wiki_graph
}
\description{
an example graph obtained from wiki to test the functionality of the second
algorithm
}
\keyword{datasets}
